requirements.txt
# HTTP
httpx==0.27.0
websockets==12.0

# Data / Indicators
pandas>=2.3.2
pandas-ta==0.4.71b0

# Scheduling
APScheduler==3.10.4
holidays==0.53

# Database
sqlalchemy>=2.0.36
aiosqlite==0.20.0

# API Server
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Telegram
python-telegram-bot==21.3

# AI (Gemini)
google-generativeai==0.7.2
google-genai>=1.10.0

# News Crawling
beautifulsoup4==4.12.3
feedparser==6.0.11

# Config
pydantic-settings==2.3.1
python-dotenv==1.0.1
